Symbolic mutable maps and structs
Previously, we had an awkward balance of immutable types for primitive
values (ints, bools) and mutable types that were restricted to storing
primitive types.  The new map and struct types naturally bridge the
gap between Python's mutable land and syntax and Z3's immutable values
by providing maps and structs that can be composed with each other and
with immutable types (so you can have, e.g., a map containing structs
with a field that is a map containing ints).

This works by introducing the notion of a "region", which is an
N-dimensional mutable array that can store primitive values and be
indexed by tuples of primitive value.  Compound types work by
recursively breaking down the compound into primitive types and
constructing regions for each primitive.