Make Symbolic the base class of all symbolic types
Now that SymbolicConst exists for symbolic constants and MetaZ3Wrapper
handles all dynamics of wrapping Z3 reference types, Symbolic can form
the base class for all symbolic types, whether or not they are
directly represented in Z3.  This centralizes the implementation of
'any' and the abstract definitions of _make_region and _select.