Rewrite SDict as a symbolic struct
The code is much shorter and more readable and the benefits are
largely the same as for SList.  This also eliminates the manual
wrapping/unwrapping when class Fs uses quantifiers and lessens the
need for class Fs to reach into the SDict structure (it doesn't
completely eliminate it for a subtle reason documented in the code;
alas).