Only check path conditions when we're done replaying
Previously, the schedule only recorded non-determinstic branches,
which meant that branches always had to invoke the solver, even during
the replay, just to determine if they were deterministic or not,
before consulting the schedule to see which direction of a
non-deterministic branch should be taken.  This is a strange thing to
do, since during replay, the solver will always come to the same
conclusions.  This patch changes this logic to record all branches in
the schedule, even if they're deterministic.  This means we can always
consult the schedule first and only need to invoke the solver to
extend the schedule when we've reached the end of replay.  In addition
to making this logic cleaner and faster, this restructuring also sets
us up to make uncheckable path conditions non-fatal.