Make Z3 types parallel Python struct/map types
Previously, we rearranged the order of struct/map type indexing so
that a given mutable struct/map was always represented by one or more
multidimsional Z3 arrays mapping to primitive type.  This was supposed
to make struct manipulation expressions simpler because we wouldn't
have to rebuild a struct every time we changed a field in it, but in
practice it significantly complicated the code and made certain basic
operations like struct assignment nearly impossible to implement.

This changes the struct and map types to instead create Z3 ADTs and
map types that directly reflect the Python type structure, which lets
us express most operations naturally and directly in Z3.

One downside of this is that the FS model now encounters various
undecidable fragments of the Z3 quantifier logic, even though the
generated expressions are clearly equivalent.  According to the Z3
authors, we were just getting "lucky" before.

This does not update the FS test generator.