"Reboot" Z3 when check returns unknown
According to the Z3 authors [1], assertion stack operations actually
affect how Z3 internally represents formulas internally, so it's
possible for check to return unknown on a solver that's used the
stack, but to succeed on a fresh solver.  So, when we get unknown, try
re-solving the assertions in a fresh solver.  This fixes almost all of
our uncheckable path constraints.

[1] http://z3.codeplex.com/workitem/26