spec.py: checkpoint rewriting the isomorphism checker
The code in same_assignments() needs to construct a condition for
two assignments being isomorphic to each other.  This is interesting
for uninterpreted sorts, where we don't care about the specific value
assignment from Z3, and care only about whether the equality pattern looks
the same.  This is made more complicated by the fact that uninterpreted
sorts show up all over the place: as values of a variable, as values in
an array, as keys in an array, as default 'else' values in an array, etc.

The tentative plan is to replace each constant in the model with a new
variable (uwrap_*), and remember the corresponding constant for later.
We first add the model's assignments, with all of the constants replaced
with uwrap variables, as an assertion.  We then construct an isomorphism
pattern for these variables, based on the equality / non-equality of
the corresponding constants, and add the NOT of this constrant as an
assertion, to get Z3 to give us a non-isomorphic model.

The current code sort-of implements this plan (in the IsomorphicMatch
class), but needs a bit more cleanup.