fs.py: use @model.methodwrap for open
This also creates anonymous values for open's ialloc and return fd,
rather than assigning them fixed names.  I think this is more correct,
since it does not constrain the ialloc or return fd choice to be the
same in all paths, but it does slow down Z3 a bit more.