mmap/munmap/mprotect/mem_read/mem_write model
One problem with our current symbolic data structure plan is that, by
using Z3's ADTs, we force Z3 to give an assignment to struct fields
that are irrelevant to our path condition, and are never accessed.
We (implicitly) rely on Z3 not generating an assignment to variables that
do not show up in constraints to help us construct compact test cases.
As a result, model enumeration generates many more cases than necessary.