Pass concrete values through MetaZ3Wrapper._wrap
Sometimes we want to use a specific wrapper type to wrap a value that
might actually be a concrete Python value of the corresponding Python
type.  Previously we only supported this for type-generic wrap and
would fail if we tried a type-specific _wrap.  Augment the classes
that have corresponding Python types so that type-specific _wrap will
pass through appropriate concrete values accordingly.

It might not be possible to trigger this right now, but it will be as
soon as we start managing our own compound values for structs and
maps, since a Python value may pass through these completely untouched
by Z3.