Never use existing solver in simsym.check
Previously, simsym.check would use the existing solver if there was
one and otherwise create a new solver.  This made for an inconsistent
interface that would account for the current path condition if there
was one, but check the expression in isolation if used outside
symbolic execution.  Since we never intentionally use this function
during symbolic execution, made it check satisfiability in isolation.

This fixes a bug in mmap/mmap that caused it to infinitely enumerate
models after the switch to the generator symbolic_apply caused the
model enumerator to use the existing solver instead of a fresh one.