fs.open: Fix time assumptions for creat path
We need to make sure time moves forward in case the creat path of open
reuses a deleted inode, but previously we made these assumptions
*after* installing the new SInode object, which would have overwritten
the existing inode's times with completely unconstrained times.  Move
these assumptions to before we install the new SInode so they apply to
the existing deleted inode's time values.