symtypes: Method to initialize a key in a map
Currently, we often create a dummy symbolic struct just so we have
something to assign to a map entry to create that entry.  While this
is a natural Python approach, creating the dummy struct introduces
extra symbolic variables that we have to deal with later.  Instead, in
most of these cases, it makes sense to simply mark the map key as
valid and then assign to the struct at that key.  The new 'create'
method facilitates this.