fs: Eliminate dummy symbolic variables
We used to have restrictions on the assignment of deeply nested struct
fields and, as a result, there were several places in the fs model
that would first create a dummy struct, set it up, and then assign the
whole struct into a symbolic map or symbolic dictionary.  This is no
longer a problem, so replace all of these dummy variables with direct
initialization of symbolic structs in maps and dictionaries.
Interestingly, not having these dummy variables and related logic
hanging around eliminates a handful of code paths from open and mmap
and slightly reduces the number of test cases generated.