simsym: Abstraction for working with models
This provides an abstraction for working with models in a natural way.
SymbolicApplyResults can provide a simsym.Model object for their code
path, which in turn can be indexed using the same variables names
provided by the user to create Symbolic variables (this is tricky
because these may not be the same as the underlying Z3 constant
names).  The returned values re-use the same Symbolic type hierarchy
that was used during symbolic execution, which enables map indexing
and struct field lookups like expected, but everything is interpreted
under the appropriate Z3 model, so these values are rooted in concrete
Python values.

This also paves the way for test generation-driven isomorphism
construction, since all concrete value interpretation is done through
simsym.Model._eval.