simsym: Rename Symbolic.any to Symbolic.var
'any' was always somewhat misleading because it could return an value
backed by an already constrained symbolic constant if a name was
reused.  It became especially problematic with the addition of the
'model' argument because, if a model was provided, the returned value
was, in fact, concrete in that model, and could not have any value.
Hence, rename the method to 'var', which better captures that it
returns a variable (or what Z3 confusingly calls a constant).