spec: Better progress reporting
The previous progress reports moved the cursor up a line to overwrite
the previous progress report, but if we had printed something else
since the last progress report, this would overwrite the last line of
that.  Instead, put the cursor in the wrap-around column after
printing the progress report so that we can still back it up to
overwrite the current progress report, but anything else written to
the terminal will cause it to wrap and start on the next line.