simsym: Use subclassing for synonym types
Previously, synonym types existed in the type hierarchy, but there was
never an *instance* of a synonym type because it was always resolved
to its base type before the value of that type was created.
Unfortunately, this meant if you had a Symbolic instance, you couldn't
map it back to a synonym type.  This was going to be a problem when we
start tracking assignments used by testgen because we won't know the
synonym types of those expressions, so we wouldn't be able to map them
back to isomorphism types.

This fixes that by directly subclassing synonym types from their base
types.