fs: Force SData length to be <= 16 DataBytes
Previously, we capped the length of files in testgen, but this caused
problems with related values such as offsets, which Z3 had constrained
relative to what it thought the file length was (which was sometimes
very large).

Now that we have a better and working assumptions system, move this
cap into an assumption on the SData type in the fs model.