simsym: Move symbolic_type into SymbolicApplyResult
Since this is really within the environment of a particular code path,
it makes more sense for it to be exposed as part of the
SymbolicApplyResult, rather than having it apply to some global state
that happens to be whatever the last symbolic_apply left behind.  This
also lets us reset constTypes on every code path, which is more
appropriate than accumulating it across code paths.