simsym: Avoid symbolic execution in SStructBase.var during model evaluation
Previously, we created symbolic field value using the field type's var
method and then immediately unwrapping it.  During model evaluation,
to avoid early evaluation of field values, we had to *not* pass the
model in to this.  However, if the field type had assumptions, that
meant it would try to assert those assumptions, which resulting in
symbolic execution during model evaluation (which occurs outside of
the symbolic context).

Fix this by not taking the shortcut through the field type's var
method.  Now the method does its own descent into the symbolic type,
much like the Symbolic.var method it overrides, with some additional
logic for the caller-specified field values.