spec: Work around non-deterministic model evaluation failure
Beginning with commit
  simsym: Avoid symbolic execution in SStructBase.var during model evaluation
Z3 started very occasionally and non-deterministically producing
models containing applications of 'array-ext'.  This is never supposed
to appear in a model and these applications interfere with model
evaluation.  In particular, we fail to evaluate
Fs.i_map.data._len[SInum!val!0] in the open/open callset (there may be
other occurrences, too).  Interestingly, this type has assumptions and
before the above commit they were being redundantly added to the
solver after symbolic execution and during model evaluation.  It seems
that this somehow nudged the solver just enough to avoid this bug.

Since this is non-deterministic, rare, and easy to detect, this patch
simply re-invokes simsym.check up to 10 times to get a clean model.