fs: Treat VMA offset as pages, not bytes
Since we can only map in pages, it doesn't make sense to record a byte
offset.  This also means that, if an iread returns no data, we now
definitely expected a SIGSEGV.  Previously, this could was incorrect
if the mapped file had length 0 because that *would* cause a SIGSEGV
(but other reads past the end of the file wouldn't).