fs_testgen: Use protection flags when setting up mappings
Previously, we created all initial mappings as read/write, which
produced incorrect results if they were supposed to be read-only.
Curiously, we already computed the protection flags string, we just
didn't use it.  With this change, we also have to be careful that we
don't then write to the page to initialize it if it's read-only.