fs_testgen: Deduplicate redundant function bodies
Many functions in the generated code have identical function bodies,
so this replaces duplicate function bodies with calls to the first
function with that body.  This roughly halves the size of the
generated code, while still keeping it easy to match up with test
output and the model file.

Without this, the fstest binary is 12MB, which is over xv6's 8MB
on-disk file size limit.  With this, it's 6MB.