fs: Remove 'nread' from SPipe
Having both nread and data be symbolic was too many degrees of
freedom.  nread also wasn't always properly constrained (in some
situations, it was greater than data.len()).  Remove nread and favor
of "shifting" data on read.  The implementation turns out to be
similar, but the shift count is always concrete and always starts at
0.