read: return EOF when no more pipe writers
Still, testgen can output incorrect code, as follows.  Suppose a test
case involves a read from a pipe that returns EAGAIN (because there
still exist writers).  The model code ensures that there exists some fd
(in either proc0 or proc1) that maps to the writer end of the pipe,
but this fd might not be referenced by anything else.  Consequently,
fstest never accesses this fd in the model, doesn't dup2() the write end
of the pipe to any fd (in any process), and the write end gets closed off.

One fix would be to capture the fd which matched the "exists" clause
in Fs.read (and which of proc0/proc1 it was from), and get testgen to
properly allocate that fd during setup.