simsym: Isolate execution graph construction code
Previously, the execution graph was tracked in the SchedNodes, which
made anything that dealt with SchedNodes or the schedule complex and
messy.  Move all of this logic into the graph generation code itself
by adding completed schedules to the graph object and letting it
construct the appropriate nodes and edges.  This makes it so we don't
have to muck with GraphNode objects all over the place, and we don't
need the funny root schedule entry that gave the graph a root but was
otherwise ignored.