simsym: Support retrieving filtered path conditions
Add an API to enable fetching the path condition with or without
assumptions and deterministic branches.  This now constructs the path
condition directly from the schedule, instead of pulling it out of the
assumptions in the solver (which worked, but always felt a little
sketchy, and gave us no way to distinguish assumptions or
deterministic branches).