use higher FDs in fs_testgen to reduce conflicts
One source of problematic test cases is pipe, which returns (two) FDs.
The spec can reason about the pipe's return FD values, and the model
specifies particular values for them, but our test code cannot control
which FD values are returned by the implementation.  As a result, some
testgen-chosen FD values (e.g., a concurrent fstat of a closed fd) may or
may not happen to be the same as some runtime-chosen FD values (e.g., pipe
fds), independent of what the model assignment required for commutativity.

The same problem can in theory show up for open(), but it probably did
not because the runtime-chosen FDs went up to 4 with open, and got up
to 5 with pipe.

Having testgen-chosen FD values start at 10 reduces this possibility.