par-spec: Fix funny temporary file names
Previously, this mutated the global args object to add per-shard
numbers to the file names.  This worked if no pool worker was ever
reused, but if a worker was reused, it would extend the
already-extended file name, resulting in names like
"test4.out.068.101.107.117.139.148".  Fix this by copying the global
args and mutating the copy.