par-mscan: Direct mscan output to a file to avoid blocking
Previously, we had mscan output to a pipe, but that would block mscan
until the Python combining code got around to reading from it.  By the
time mscan started writing, it was mostly done, but not entirely done,
so this limited par-mscan's parallelism at the end.  Instead, redirect
its output to a file to avoid blocking mscan.  This noticeably
improves the performance of the tail of par-mscan.