simsym: Fix struct equality
Previously, we flattened the compound values of both operands and then
compared these flattened values, but flattening is non-deterministic,
so the two flattened compounds were not guaranteed to put their fields
in the same order.  Somehow this mostly worked, except for atime and
nlink tended to get mixed up in Fs equality check (which also happened
to have the same underlying Z3 type, which is why this went undetected
for so long).

Fix this by parallel mapping across both compounds to construct a
compound of equality expressions, and only *then* flatten and join the
equality expressions.