simsym: Avoid identity map updates
This looks like a goofy optimization, but once you peel back all of
the layers, it turns out it applies more often than it doesn't apply
and can reduces many-page expressions to just a few lines.

Previously, any time we updated a field in a struct in an array, we
would wrap *every* field in the struct in a new Store expression; for
all of the fields that didn't change, these Stores did nothing.  Now
we check for this case and avoid making the Store expression.