spec: Compute and record idempotence of callsets
This is intended to help us classify the "hard to scale" cases in xv6.
There are still some hiccups: its turns out that two open(O_CREAT)s of
the same file name in different processes are *not* idempotent by the
definition used here because they both modify the local process' FD
table, and hence affect state, even though *part* of the open
operation is an idempotent update.