fs_testgen: Deduplicate whole functions
Previously we only deduplicated the function body and replaced it with
a call to the source function.  This makes splitting a testgen.c file
into shards really hard.  So, instead, don't emit the function at all
(except as a comment) and just use the source function's name in the
tests array.