model, fs: Declare internal vars directly, rather than as arguments
Previously, any time we needed an internal variable, we had to declare
it as an argument that would get instantiated by model.methodwrap.
This was necessary because only model.methodwrap knew the current
sequence name, so only it could construct variable names that were
distinct for each call set sequence.  Now that we just use anonymous
names for these, it's much simpler to declare them where we need them
in fs and rip out the weird and confusing "internal argument"
machinery from model.methodwrap.