spec: Compute projected idempotence
Our current whole-state idempotence test turns out not to be exactly
what we want because many pairs of operations are idempotent only for
the part of the state they conflict on, but also update other distinct
parts of the state.  For example, open O_TRUNCs perform idempotent
updates on the file length, but also modify distinct per-process FD
tables.

This adds a test for finer-grained idempotence that projects out every
node in the state's type tree and determines for which projections an
operation is idempotent.