fs_testgen: Add a DynamicDict to map SOffsets to physical offsets
Unlike the other DynamicDict-mapped types, SOffsets already have
physical meaning.  However, they're off from physical offsets by a
multiplicative factor, so this keeps that knowledge in one place.

The real point of this is that we treat SOffset as uninterpreted for
isomophrism generation and this will provide the semantic information
we need to isolate equality constraints to only offsets within the
same inode.