fs: Make VMA offsets be in dataval units, not page units
Previously, VMA offsets were in pages and we applied a translation
wherever we needed it.  Besides being ugly, this is going to cause
grief when we start tracking offset usage in fs_testgen, so make VMA
offsets also be in dataval units and instead require that they have
the right alignment.