simsym/fs_testgen: Explicitly fetch concrete values
Previously, during concolic execution of testgen, we automatically
evaluated symbolic values to concrete values as soon as we reached a
leaf value.  However, this loses a lot of semantic information, since
there's no way to access the final symbolic expression, and also gives
us little control over what expressions are recorded in the model's
assignment history.

This changes simsym to delay concrete evaluation until the caller
accesses new properties "val" or "someval" and changes fs_testgen to
use these properties to retrieve concrete values explicitly when they
are needed.  This is somewhat simplified by the introduction of a new
"Interpreter" collection that acts like a DynamicDict, but takes
unevaluated symbolic expressions and implicitly forces them to
concrete values.

This slightly changes the set of generated tests, largely because we
often don't force things like map indexes to concrete values.  It
increases the number of tests for some callsets and decreases it for
others, but works out to ~3500 more tests in the end.

The ultimate goal of this is to keep around symbolic information
longer so we can do more precise isomorphism generation.