symtypes: Correct equality for lists
Previously, symbolic lists relied on the default structural equality,
but this is incorrect for a variety of reasons: it ignores the meaning
of _start, so equal lists may appear unequal; and it considers
elements outside the bounds of the list, so lists that were equal
could appear unequal if these didn't agree, or even appear both equal
and unequal if these elements were unconstrained.

Ultimately, this has little effect on the FS model because its lists
are so unconstrained (so the reason for equality/inequality may have
been wrong, but the conclusion wasn't).