symtypes: Correct equality for dicts
Previously, dicts could have been considered unequal because of the
values of invalid keys.  Fix this by defining a correct equality
method.

Again, this has little (or no) effect on the FS model because it uses
unconstrained dicts, so Z3 could already find some reason for them to
be unequal (though it may have been the wrong reason).