fs_testgen: Create extra pipe FDs when necessary to keep pipes open
This fixes a bug where pipes would sometimes get closed when they
should remain open.  Since testgen would only open the pipe FDs that
were directly needed by the test, it could miss other open pipe ends
that were implied by the path condition.  For example, closing the
write end of a pipe and reading the read end of a pipe commute if
there is another FD for the write end of the pipe that remains open,
but since that other write end FD never appeared in a test, the
generated code would never open it.  We fix this by specifically
testing for the existence of pipe FDs beyond the ones testgen already
knows about for every pipe end.

This makes all result and commutativity sanity checks (fstest -r and
-c) pass on Linux for {read,write}/close.