spec: Make projected idempotent per-test instead of per-path
The per-path idempotence only told us if there was some satisfying
assignment of the path condition that was idempotent.  But that could
mean that just one test of that path was idempotent and the rest
weren't (for example, lseek_lseek_pad0).  Make idempotence per-test by
constraining the path condition with each test's isomorphism condition.