simsym: Fix handling of unknown/unsat and unsat/unknown
Previously, if evaluating an SBool gave unknown/unsat or
unsat/unknown, we would still fork the schedule and the unsat
direction would be treated like unknown.  This manifested as
UncheckableConstraintErrors with a reason of "ok", though we would
stop at the exception, so we never tried to explore down unsatisfiable
paths.