spec: Implement full SIM commutativity test
Previously, we effectively tested for SI commutativity, since we only
compared all final results.  This was equivalent to SIM commutativity
for pairs, which was all we needed for the paper, but omitted the
intermediate state comparisons necessary to test for monotonicity.

This implements those intermediate state comparisons using a new
recursive algorithm designed to perform the minimum number of
comparisons possible using a walk over an n-dimensional cube.

Even for pairs, this winds up exploring fewer non-commutative code
paths because it bail out earlier on result divergence (we, of course,
explore the same number of *commutative* code paths).

This slightly changes the 'diverge' field of the model.out file.  ''
still means commutativity (so existing code doesn't break), but
non-empty strings are more informative about which calls in which
sequences diverged.