viewer: Interactive web-based browser for Commuter output
This initial version displays heatmaps of mscan data, separated by OS
(like in the paper), and a table of test cases.  Clicking on a cell in
the heat map filters the data to matched test cases in that cell,
which updates the table.  Clicking a row in the table expands it to
show the full JSON record from mscan.