viewer: Split databases into general and sharing details
This lets us load the small general database (~120K for each of Linux
and sv6, gzip-able down to 10K-20K) that has enough information to
show the heatmap and table, but pushes the detailed sharing
information into a separate database that can be loaded if necessary.

At the moment we don't load the sharing details database.