viewer: Compute table columns before building table
Previously, the table renderer used a single pass over the database
and dynamically added columns as it came across them.  This was both a
pain and slightly buggy in some cases where we needed to know the
total number of columns before the scan was necessarily complete.

Replace this with a much simpler two-pass algorithm that collects
columns and then builds the table.