viewer: Split test structure and results databases
This generates a "template" database containing the common test
structure shared between different runs of the same testgen and
between the main and sharing databases for the same run.  This shrinks
the Linux and sv6 main databases to 1.2K and 22K (uncompressed), plus
a shared 133K (12K compressed) template database.