viewer: Add "heat bar" between heatmap and table
This visualizes the relative proportion of conflict-free and
conflicted tests for the selection and gives summary statistics in
both absolute numbers and percentages.  It also offers a convenient
place to filter to only conflict-free or conflicted cases.

One problem with the current implementation is that there's no way to
select everything once you've selected either conflict-free or
conflicted (unlike the heat map, there's nowhere else on the canvas to
click, not that that's a good interface for the heatmap).