simsym: Correct and robust Symbolic.eq
Previously, Symbolic.eq assumed that compound dictionaries would have
the same keys and would iterate in the same order and didn't properly
handle compounds with concrete values.  In practice, it was never
called with anything that would violate these, but this depended on
subtle aspects of our implementation and CPython itself.  Fix these
problems to make Symbolic.eq robust.